+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to the "vscode-rufo" extension will be documented in this fi
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
-- Format Ruby source code via Rufo
+
+- Show error messages from Rufo
 
 ## [0.0.3] - 2020-12-21